+++ react/IconUser.js
@@ -1,0 +1,3 @@
+import User from './components/icon/User/index'
+
+export default User